+++ poc/native_audio_engine_ios/ios/Runner/AppDelegate.swift
@@ -1,0 +1,53 @@
+import Flutter
+import UIKit
+
+@main
+@objc class AppDelegate: FlutterAppDelegate, FlutterImplicitEngineDelegate {
+    private let audioEngine = AudioEngine()
+    private var nativeChannel: FlutterMethodChannel?
+
+    override func application(
+        _ application: UIApplication,
+        didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
+    ) -> Bool {
+        return super.application(application, didFinishLaunchingWithOptions: launchOptions)
+    }
+
+    func didInitializeImplicitFlutterEngine(_ engineBridge: FlutterImplicitEngineBridge) {
+        GeneratedPluginRegistrant.register(with: engineBridge.pluginRegistry)
+
+        let messenger = engineBridge.applicationRegistrar.messenger()
+        let channel = FlutterMethodChannel(
+            name: "com.synchorus.poc/native_audio",
+            binaryMessenger: messenger
+        )
+        self.nativeChannel = channel
+
+        channel.setMethodCallHandler { [weak self] call, result in
+            guard let self = self else { return }
+            switch call.method {
+            case "start":
+                result(self.audioEngine.start())
+            case "stop":
+                result(self.audioEngine.stop())
+            case "getTimestamp":
+                result(self.audioEngine.getTimestamp())
+            case "seekToFrame":
+                guard let args = call.arguments as? [String: Any],
+                      let newFrame = (args["newFrame"] as? NSNumber)?.int64Value
+                else {
+                    result(
+                        FlutterError(
+                            code: "BAD_ARGS", message: "newFrame required",
+                            details: nil))
+                    return
+                }
+                result(self.audioEngine.seekToFrame(newFrame))
+            case "getVirtualFrame":
+                result(self.audioEngine.getVirtualFrame())
+            default:
+                result(FlutterMethodNotImplemented)
+            }
+        }
+    }
+}